+++ integration/querier_sharding_test.go
@@ -55,7 +55,8 @@ func TestQuerierNoShardingWithQueryScheduler(t *testing.T) {
 
 func runQuerierShardingTest(t *testing.T, cfg querierShardingTestConfig) {
 	// Going to high starts hitting file descriptor limit, since we run all queriers concurrently.
-	const numQueries = 100
+	const batchSize = 100
+	const numQueries = 500
 
 	s, err := e2e.NewScenario(networkName)
 	require.NoError(t, err)
@@ -141,25 +142,29 @@ func runQuerierShardingTest(t *testing.T, cfg querierShardingTestConfig) {
 		require.NoError(t, queryFrontend.WaitSumMetrics(e2e.Equals(2), "cortex_query_frontend_connected_clients"))
 	}
 
+	batches := generateBatches(batchSize, numQueries)
+
 	wg := sync.WaitGroup{}
 
 	// Run all queries concurrently to get better distribution of requests between queriers.
-	for i := 0; i < numQueries; i++ {
-		wg.Add(1)
+	for _, concurrentQueries := range batches {
+		for i := 0; i < concurrentQueries; i++ {
+			wg.Add(1)
 
-		go func() {
-			defer wg.Done()
-			c, err := e2ecortex.NewClient("", queryFrontend.HTTPEndpoint(), "", "", userID)
-			require.NoError(t, err)
+			go func() {
+				defer wg.Done()
+				c, err := e2ecortex.NewClient("", queryFrontend.HTTPEndpoint(), "", "", userID)
+				require.NoError(t, err)
 
-			result, err := c.Query("series_1", now)
-			require.NoError(t, err)
-			require.Equal(t, model.ValVector, result.Type())
-			assert.Equal(t, expectedVector, result.(model.Vector))
-		}()
+				result, err := c.Query("series_1", now)
+				require.NoError(t, err)
+				require.Equal(t, model.ValVector, result.Type())
+				assert.Equal(t, expectedVector, result.(model.Vector))
+			}()
+		}
+
+		wg.Wait()
 	}
-
-	wg.Wait()
 
 	require.NoError(t, queryFrontend.WaitSumMetrics(e2e.Equals(numQueries), "cortex_query_frontend_queries_total"))
 
@@ -193,4 +198,20 @@ func runQuerierShardingTest(t *testing.T, cfg querierShardingTestConfig) {
 	assertServiceMetricsPrefixes(t, Querier, querier2)
 	assertServiceMetricsPrefixes(t, QueryFrontend, queryFrontend)
 	assertServiceMetricsPrefixes(t, QueryScheduler, queryScheduler)
+}
+
+func generateBatches(batchSize, total int) []int {
+	remain := total
+	batches := []int{}
+	for remain != 0 {
+		if remain > batchSize {
+			batches = append(batches, batchSize)
+			remain -= batchSize
+		} else {
+			batches = append(batches, remain)
+			remain = 0
+		}
+	}
+
+	return batches
 }